move styles to organism instead of page
